requirements.txt
numpy==1.24.2
pandas==2.0.0
bokeh==2.4.3
jupyterlab==3.5.3
holoviews==1.15.3
requests==2.28.2
openpyxl==3.1.2
biopython==1.81
networkx==3.1
scipy==1.10.1
matplotlib